fix(8.10): keep zoned migration resources release-local
