+++ charts/camunda-platform-8.10/templates/orchestration/configmap.yaml
@@ -1,3 +1,3 @@
 {{- if eq (include "camundaPlatform.orchestrationEnabled" .) "true" -}}
-{{- include "orchestration.renderZoneAware" (dict "manifest" "orchestration.configmapManifest" "context" .) -}}
+{{- include "orchestration.renderBrokerGenerations" (dict "manifest" "orchestration.configmapManifest" "context" .) -}}
 {{- end -}}
