+++ charts/camunda-platform-8.10/templates/orchestration/files/_application.yaml
@@ -89,24 +89,23 @@ camunda:
         {{- end }}
     {{- end }}
     # zeebe.broker.cluster
-    {{- /* NOTE: zoned brokers all live in-cluster (the single-cluster multi-AZ case this
-         mode targets), so every zone's headless Service is addressable and the full list
-         can be generated. A zoned release scoped to one cross-cluster zone (dual-region)
-         cannot resolve the other regions' zones from in-cluster DNS; override the whole
-         list via CAMUNDA_CLUSTER_INITIALCONTACTPOINTS in that case, same escape hatch as
-         legacy multi-region. */ -}}
     {{- if eq (include "orchestration.zoned" .) "true" }}
+    {{- if or $mr.keepUnzonedBrokers (ne (include "camundaPlatform.multiregionSpread" $) "true") }}
     initial-contact-points:
-    {{- range $mr.zones }}
-    {{- $zoneFullname := include "orchestration.zoneFullname" (dict "context" $ "zone" .name) }}
-    {{- range untilStep 0 (int .numberOfBrokers) 1 }}
+    {{- if ne (include "camundaPlatform.multiregionSpread" $) "true" }}
+    {{- $zoneFullname := include "orchestration.zoneFullname" (dict "context" $ "zone" $mr.zone) }}
+    {{- range untilStep 0 (int (include "orchestration.zoneBrokers" $)) 1 }}
       - {{ $zoneFullname }}-{{ . }}.{{ $zoneFullname }}:{{ $.Values.orchestration.service.internalPort }}
     {{- end }}
     {{- end }}
     {{- if $mr.keepUnzonedBrokers }}
-    {{- range untilStep 0 (int $.Values.orchestration.clusterSize) 1 }}
+    {{- range untilStep 0 (int (include "orchestration.numberedReplicas" $)) 1 }}
       - {{ include "orchestration.fullname" $ }}-{{ . }}.{{ include "orchestration.fullname" $ }}:{{ $.Values.orchestration.service.internalPort }}
     {{- end }}
+    {{- end }}
+    {{- else }}
+    # Multi-zone deployments: initial-contact-points must be provided manually via environment variable.
+    # Set CAMUNDA_CLUSTER_INITIALCONTACTPOINTS with fully-qualified domain names for all brokers across zones.
     {{- end }}
     {{- else if ne (include "camundaPlatform.multiregionSpread" $) "true" }}
     initial-contact-points:
