+++ charts/camunda-platform-8.10/templates/orchestration/poddisruptionbudget.yaml
@@ -1,3 +1,3 @@
 {{- if and (eq (include "camundaPlatform.orchestrationEnabled" .) "true") .Values.orchestration.podDisruptionBudget.enabled -}}
-{{- include "orchestration.renderZoneAware" (dict "manifest" "orchestration.podDisruptionBudgetManifest" "context" .) -}}
+{{- include "orchestration.renderBrokerGenerations" (dict "manifest" "orchestration.podDisruptionBudgetManifest" "context" .) -}}
 {{- end -}}
