+++ charts/camunda-platform-8.10/templates/orchestration/service-headless.yaml
@@ -2,5 +2,5 @@
 The headless service is used for Zeebe brokers Pod-to-Pod communication.
 */}}
 {{- if eq (include "camundaPlatform.orchestrationEnabled" .) "true" -}}
-{{- include "orchestration.renderZoneAware" (dict "manifest" "orchestration.serviceHeadless" "context" . "renderUnzoned" true) -}}
+{{- include "orchestration.renderHeadlessServices" (dict "context" .) -}}
 {{- end -}}
