+++ charts/camunda-platform-8.10/templates/orchestration/serviceaccount.yaml
@@ -1,3 +1,21 @@
 {{- if and (eq (include "camundaPlatform.orchestrationEnabled" .) "true") .Values.orchestration.serviceAccount.enabled -}}
-{{- include "orchestration.renderZoneAware" (dict "manifest" "orchestration.serviceAccountManifest" "context" .) -}}
-{{- end -}}
+apiVersion: v1
+kind: ServiceAccount
+metadata:
+  name: {{ include "orchestration.serviceAccountName" . }}
+  labels:
+    {{- include "common.tplvalues.merge-overwrite" (dict
+      "values" (list
+        (include "orchestration.labels" .)
+        .Values.orchestration.serviceAccount.labels
+      )
+      "context" .
+    ) | nindent 4 }}
+  {{- $secretStoreAnnotations := fromYaml (include "camundaPlatform.secretStore.serviceAccountAnnotations" .) }}
+  {{- $annotations := mergeOverwrite (deepCopy (.Values.orchestration.serviceAccount.annotations | default dict)) $secretStoreAnnotations }}
+  {{- with $annotations }}
+  annotations:
+    {{- toYaml . | nindent 4 }}
+  {{- end }}
+automountServiceAccountToken: {{ .Values.orchestration.serviceAccount.automountServiceAccountToken }}
+{{- end }}
