+++ charts/camunda-platform-8.10/templates/orchestration/statefulset.yaml
@@ -1,3 +1,3 @@
 {{- if eq (include "camundaPlatform.orchestrationEnabled" .) "true" -}}
-{{- include "orchestration.renderZoneAware" (dict "manifest" "orchestration.statefulset" "context" .) -}}
+{{- include "orchestration.renderBrokerGenerations" (dict "manifest" "orchestration.statefulset" "context" .) -}}
 {{- end -}}
