+++ charts/camunda-platform-8.10/test/unit/orchestration/golden/migration-configmap.golden.yaml
@@ -97,12 +97,7 @@ data:
                 priority: 80
         # zeebe.broker.cluster
         initial-contact-points:
-          - camunda-platform-test-zeebe-zone-a-0.camunda-platform-test-zeebe-zone-a:26502
-          - camunda-platform-test-zeebe-zone-b-0.camunda-platform-test-zeebe-zone-b:26502
-          - camunda-platform-test-zeebe-zone-c-0.camunda-platform-test-zeebe-zone-c:26502
           - camunda-platform-test-zeebe-0.camunda-platform-test-zeebe:26502
-          - camunda-platform-test-zeebe-1.camunda-platform-test-zeebe:26502
-          - camunda-platform-test-zeebe-2.camunda-platform-test-zeebe:26502
     
       api:
         grpc:
