+++ charts/camunda-platform-8.10/test/unit/orchestration/golden/migration-serviceaccount.golden.yaml
@@ -3,24 +3,6 @@
 apiVersion: v1
 kind: ServiceAccount
 metadata:
-  name: camunda-platform-test-zeebe-zone-a
-  labels:
-    app: camunda-platform
-    app.kubernetes.io/component: zeebe-broker
-    app.kubernetes.io/instance: camunda-platform-test
-    app.kubernetes.io/managed-by: Helm
-    app.kubernetes.io/name: camunda-platform
-    app.kubernetes.io/part-of: camunda-platform
-    app.kubernetes.io/version: 8.10.0-alpha5
-    camunda.io/zone: zone-a
-automountServiceAccountToken: false
-
-
----
-# Source: camunda-platform/templates/orchestration/serviceaccount.yaml
-apiVersion: v1
-kind: ServiceAccount
-metadata:
   name: camunda-platform-test-zeebe
   labels:
     app: camunda-platform
