+++ charts/camunda-platform-8.10/test/unit/orchestration/golden/zoned-configmap.golden.yaml
@@ -96,10 +96,8 @@ data:
                 number-of-replicas: 1
                 priority: 80
         # zeebe.broker.cluster
-        initial-contact-points:
-          - camunda-platform-test-zeebe-zone-a-0.camunda-platform-test-zeebe-zone-a:26502
-          - camunda-platform-test-zeebe-zone-b-0.camunda-platform-test-zeebe-zone-b:26502
-          - camunda-platform-test-zeebe-zone-c-0.camunda-platform-test-zeebe-zone-c:26502
+        # Multi-zone deployments: initial-contact-points must be provided manually via environment variable.
+        # Set CAMUNDA_CLUSTER_INITIALCONTACTPOINTS with fully-qualified domain names for all brokers across zones.
     
       api:
         grpc:
