+++ charts/camunda-platform-8.10/test/unit/orchestration/golden/zoned-serviceaccount.golden.yaml
@@ -3,7 +3,7 @@
 apiVersion: v1
 kind: ServiceAccount
 metadata:
-  name: camunda-platform-test-zeebe-zone-a
+  name: camunda-platform-test-zeebe
   labels:
     app: camunda-platform
     app.kubernetes.io/component: zeebe-broker
@@ -12,5 +12,4 @@ metadata:
     app.kubernetes.io/name: camunda-platform
     app.kubernetes.io/part-of: camunda-platform
     app.kubernetes.io/version: 8.10.0-alpha5
-    camunda.io/zone: zone-a
 automountServiceAccountToken: false
