+++ charts/camunda-platform-8.10/test/unit/orchestration/golden/zoned-statefulset.golden.yaml
@@ -138,7 +138,7 @@ spec:
           emptyDir: {}
         - name: tmp
           emptyDir: {}
-      serviceAccountName: camunda-platform-test-zeebe-zone-a
+      serviceAccountName: camunda-platform-test-zeebe
       securityContext:
         fsGroup: 1001
         runAsNonRoot: true
