+++ charts/camunda-platform-8.10/test/unit/orchestration/zone_aware_serviceaccount_test.go
@@ -1,0 +1,110 @@
+// Copyright 2026 Camunda Services GmbH
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package orchestration
+
+import (
+	"camunda-platform/test/unit/testhelpers"
+	"path/filepath"
+	"strings"
+	"testing"
+
+	"github.com/gruntwork-io/terratest/modules/random"
+	"github.com/stretchr/testify/require"
+	"github.com/stretchr/testify/suite"
+)
+
+type ZoneAwareServiceAccountTest struct {
+	suite.Suite
+	chartPath  string
+	release    string
+	namespace  string
+	templates  []string
+	valuesFile string
+}
+
+func TestZoneAwareServiceAccountTemplate(t *testing.T) {
+	t.Parallel()
+
+	chartPath, err := filepath.Abs("../../../")
+	require.NoError(t, err)
+
+	suite.Run(t, &ZoneAwareServiceAccountTest{
+		chartPath:  chartPath,
+		release:    "camunda-platform-test",
+		namespace:  "camunda-platform-" + strings.ToLower(random.UniqueId()),
+		templates:  []string{"templates/orchestration/serviceaccount.yaml"},
+		valuesFile: filepath.Join(chartPath, "test/unit/orchestration/testdata/values-zone-aware-migration.yaml"),
+	})
+}
+
+func (s *ZoneAwareServiceAccountTest) TestMigrationUsesOneServiceAccount() {
+	testCases := []testhelpers.TestCase{
+		{
+			Name: "TestDefaultName",
+			Verifier: func(t *testing.T, output string, err error) {
+				require.NoError(t, err)
+				require.Equal(t, 1, strings.Count(output, "kind: ServiceAccount"))
+				require.Contains(t, output, "name: camunda-platform-test-zeebe")
+				require.NotContains(t, output, "name: camunda-platform-test-zeebe-zone-a")
+			},
+		},
+		{
+			Name: "TestCustomName",
+			Values: map[string]string{
+				"orchestration.serviceAccount.name": "custom-sa",
+			},
+			Verifier: func(t *testing.T, output string, err error) {
+				require.NoError(t, err)
+				require.Equal(t, 1, strings.Count(output, "kind: ServiceAccount"))
+				require.Contains(t, output, "name: custom-sa")
+			},
+		},
+	}
+	for i := range testCases {
+		testCases[i].ValuesFiles = []string{s.valuesFile}
+	}
+
+	testhelpers.RunTestCasesE(s.T(), s.chartPath, s.release, s.namespace, s.templates, testCases)
+}
+
+func (s *ZoneAwareServiceAccountTest) TestBothBrokerGenerationsUseTheSharedServiceAccount() {
+	testCases := []testhelpers.TestCase{
+		{
+			Name: "TestDefaultName",
+			Verifier: func(t *testing.T, output string, err error) {
+				require.NoError(t, err)
+				require.Equal(t, 2, strings.Count(output, "serviceAccountName: camunda-platform-test-zeebe"))
+			},
+		},
+		{
+			Name: "TestCustomName",
+			Values: map[string]string{
+				"orchestration.serviceAccount.name": "custom-sa",
+			},
+			Verifier: func(t *testing.T, output string, err error) {
+				require.NoError(t, err)
+				require.Equal(t, 2, strings.Count(output, "serviceAccountName: custom-sa"))
+			},
+		},
+	}
+	for i := range testCases {
+		testCases[i].ValuesFiles = []string{s.valuesFile}
+	}
+
+	testhelpers.RunTestCasesE(
+		s.T(), s.chartPath, s.release, s.namespace,
+		[]string{"templates/orchestration/statefulset.yaml"}, testCases,
+	)
+}
